requirements.txt
amqp==2.2.2
beautifulsoup4==4.6.0
celery==4.1.0
cloudinary==1.11.0
Django==2.0.4
dlib==19.10.0
face-recognition==1.2.2
face-recognition-models==0.3.0
jsonfield==2.0.1
piplapis-python==5.1.4
requests==2.18.4
spacy==2.0.11
